adding 2 new files

+++ ft_strnstr.c
@@ -9,13 +9,14 @@
 /*   Updated: 2024/05/22 17:27:14 by jasooi           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+#include <stdio.h>
+#include <bsd/string.h>
+#include "libft.h"
 
-//#include "libft.h"
-
-char	*ft_strnstr(char *str, char *to_find, int n)
+char	*ft_strnstr(char *str, char *to_find, size_t n)
 {
-	int		i;
-	int		j;
+	size_t		i;
+	size_t		j;
 
 	i = 0;
 	if (to_find[i] == '\0')
@@ -37,20 +38,19 @@ char	*ft_strnstr(char *str, char *to_find, int n)
 }
 
 
-#include <stdio.h>
-#include <bsd/string.h>
+
 int     main(void)
 {
         char    str1[] = "your xylophone now!";
         char    str2[] = "xylop";
-	char	str3[] = "xy";
+	    char	str3[] = "xy";
         int    n = 15;
-	int	m = 7;
+	    int	m = 7;
 
         printf(" n = %d | expected outcome = %s | my fn's outcome = %s\n", n, strnstr(str1, str2, n), ft_strnstr(str1, str2, n));
 
-	printf(" n = %d | expected outcome = %s | my fn's outcome = %s\n", n, strnstr(str1, str3, n), ft_strnstr(str1, str3, n));
+	    printf(" n = %d | expected outcome = %s | my fn's outcome = %s\n", n, strnstr(str1, str3, n), ft_strnstr(str1, str3, n));
 
-	printf(" m = %d | expected outcome = %s | my fn's outcome = %s\n", m, strnstr(str1, str2, m), ft_strnstr(str1, str2, m));
+	    printf(" m = %d | expected outcome = %s | my fn's outcome = %s\n", m, strnstr(str1, str2, m), ft_strnstr(str1, str2, m));
 }
